--- notifications/src/main/java/deti/fitmonitor/notifications/NotificationsApplication.java
@@ -1,13 +1,0 @@
-package deti.fitmonitor.notifications;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NotificationsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(NotificationsApplication.class, args);
-	}
-
-}
