--- notifications/src/main/java/deti/fitmonitor/notifications/repositories/paymentsRepository.java
@@ -1,14 +1,0 @@
-package deti.fitmonitor.notifications.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-import deti.fitmonitor.notifications.models.Payments;
-
-@Repository
-public interface paymentsRepository extends JpaRepository<Payments, String> {
-    Optional<Payments> findByUserSub(String userSub);
-    
-}
