--- notifications/src/test/java/deti/fitmonitor/notifications/NotificationsApplicationTests.java
@@ -1,9 +1,0 @@
-package deti.fitmonitor.notifications;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class NotificationsApplicationTests {
-
-}
