+++ payments/src/main/java/deti/fitmonitor/payments/PaymentsApplication.java
@@ -1,0 +1,13 @@
+package deti.fitmonitor.payments;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PaymentsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PaymentsApplication.class, args);
+	}
+
+}
