+++ payments/src/main/java/deti/fitmonitor/payments/configuration/SwaggerConfig.java
@@ -1,0 +1,21 @@
+package deti.fitmonitor.payments.configuration;
+
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@EnableWebMvc
+public class SwaggerConfig implements WebMvcConfigurer{
+    @Override
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+        registry.addResourceHandler("/swagger-ui/**")
+                .addResourceLocations("classpath:/META-INF/resources/webjars/springdoc-openapi-ui/")
+                .resourceChain(false);
+
+        registry.addResourceHandler("/**")
+                .addResourceLocations("classpath:/META-INF/resources/", "classpath:/resources/", "classpath:/static/", "classpath:/public/");
+    }
+}
