+++ payments/src/main/resources/application.properties
@@ -1,0 +1,21 @@
+spring.application.name=payments
+
+# Spring Data Source Config
+spring.datasource.url=${SPRING_DATASOURCE_URL}
+spring.datasource.username=${SPRING_DATASOURCE_USERNAME}
+spring.datasource.password=${SPRING_DATASOURCE_PASSWORD}
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
+
+spring.jpa.hibernate.ddl-auto=update
+
+#AWS Cognito
+external.auth.token.url=${EXTERNAL_AUTH_TOKEN_URL}
+external.auth.client.credentials=${EXTERNAL_AUTH_CLIENT_CREDENTIALS}
+
+cognito.jwks.url=${COGNITO_JWKS_URL}
+
+server.port=8181
+
+stripe.api.key=${STRIPE_API_KEY}
+stripe.webhook.secret=${STRIPE_WEBHOOK_SECRET}
