+++ payments/src/test/java/deti/fitmonitor/payments/PaymentsApplicationTests.java
@@ -1,0 +1,8 @@
+package deti.fitmonitor.payments;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PaymentsApplicationTests {
+
+}
